Moved colour code to q_shared.c Fixed some warnings

+++ src/botlib/l_script.h
@@ -180,7 +180,7 @@ typedef struct script_s
 	char *lastscript_p;				//script pointer before reading token
 	char *whitespace_p;				//begin of the white space
 	char *endwhitespace_p;			//end of the white space
-	int length;						//length of the script in bytes
+	size_t length;					//length of the script in bytes
 	int line;						//current line in script
 	int lastline;					//line before reading token
 	int tokenavailable;				//set by UnreadLastToken
